Switched to pytest fixture
